+++ src/Html.php
@@ -5,13 +5,13 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\bootstrap;
+namespace yii\bootstrap4;
 
 /**
  * Html is an enhanced version of [[\yii\helpers\Html]] helper class dedicated to the Bootstrap needs.
  * This class inherits all functionality available at [[\yii\helpers\Html]] and can be used as substitute.
  *
- * Attention: do not confuse [[\yii\bootstrap\Html]] and [[\yii\helpers\Html]], be careful in which class
+ * Attention: do not confuse [[\yii\bootstrap4\Html]] and [[\yii\helpers\Html]], be careful in which class
  * you are using inside your views.
  *
  * @author Paul Klimov <klimov.paul@gmail.com>