+++ tests/CollapseTest.php
@@ -2,7 +2,7 @@
 namespace yiiunit\extensions\bootstrap;
 
 use yii\base\DynamicModel;
-use yii\bootstrap\Collapse;
+use yii\bootstrap4\Collapse;
 use yii\widgets\ActiveForm;
 
 /**