+++ tests/HtmlTest.php
@@ -3,7 +3,7 @@
 namespace yiiunit\extensions\bootstrap;
 
 use yii\base\DynamicModel;
-use yii\bootstrap\Html;
+use yii\bootstrap4\Html;
 
 /**
  * @group bootstrap