+++ CHANGELOG.md
@@ -4,4 +4,5 @@ Yii Framework 2 bootstrap4 extension Change Log
 1.0.0 under development
 -----------------------
 
+- (enh #3): Update namespace to `bootstrap4` for the new extension. (kartik-v)
 - Initial release (forked from yii2-bootstrap and adjusted to fit bootstrap 4)